+++ quilt-rs/src/io/remote/object/put_and_request_checksum.rs
@@ -1,6 +1,5 @@
 use std::path::Path;
 
-use aws_sdk_s3::error::DisplayErrorContext;
 use aws_sdk_s3::operation::put_object::PutObjectOutput;
 use aws_sdk_s3::primitives::ByteStream;
 use aws_sdk_s3::types::ChecksumAlgorithm;
@@ -11,6 +10,7 @@ use crate::checksum::Sha256ChunkedHash;
 use crate::error::ChecksumError;
 use crate::error::S3Error;
 use crate::error::S3ErrorKind;
+use crate::io::remote::describe_sdk_error;
 use crate::io::remote::HostChecksums;
 use crate::io::remote::HostConfig;
 use crate::uri::S3Uri;
@@ -57,7 +57,7 @@ pub async fn put_and_request_checksum(
         .map_err(|err| {
             Error::S3(S3Error {
                 host: host_config.host.clone(),
-                kind: S3ErrorKind::UploadFile(DisplayErrorContext(err).to_string()),
+                kind: S3ErrorKind::UploadFile(describe_sdk_error(err)),
             })
         })?;
     let checksum = match host_config.checksums {